FLS-1470 - Moving test-specific functions from import_config package into conftest.py, then removing now redundant import_config package
