--- app/import_config/files_to_import/Contact details template.json
@@ -1,181 +1,0 @@
-{
-    "metadata": {},
-    "startPage": "/lead-contact-details",
-    "pages": [
-        {
-            "title": "Lead contact details",
-            "path": "/lead-contact-details",
-            "components": [
-                {
-                    "name": "rILgUu",
-                    "options": {},
-                    "type": "Html",
-                    "content": "<p class=\"govuk-body\">They will receive all the information about this application.</p>",
-                    "schema": {}
-                },
-                {
-                    "name": "CgCZmL",
-                    "options": {},
-                    "type": "TextField",
-                    "title": "Lead contact full name",
-                    "schema": {}
-                },
-                {
-                    "name": "XHFsXs",
-                    "options": {},
-                    "type": "TextField",
-                    "title": "Lead contact job title",
-                    "schema": {}
-                },
-                {
-                    "name": "qSOFAl",
-                    "options": {},
-                    "type": "EmailAddressField",
-                    "title": "Lead contact email address",
-                    "schema": {}
-                },
-                {
-                    "name": "RyHYTL",
-                    "options": {},
-                    "type": "NumberField",
-                    "title": "Lead contact telephone number",
-                    "schema": {}
-                }
-            ],
-            "next": [
-                {
-                    "path": "/is-the-lead-contact-the-same-person-as-the-authorised-signatory"
-                }
-            ],
-            "section": "roYQzW"
-        },
-        {
-            "path": "/is-the-lead-contact-the-same-person-as-the-authorised-signatory",
-            "title": "Is the lead contact the same person as the authorised signatory?",
-            "components": [
-                {
-                    "name": "SDJIoS",
-                    "options": {
-                        "exposeToContext": true
-                    },
-                    "type": "YesNoField",
-                    "title": "Is the lead contact the same person as the authorised signatory?",
-                    "hint": "<p class=\"govuk-body\">An authorised signatory:</p>\n<ul class=\"govuk-list govuk-list--bullet\">\n  <li>is allowed to act on behalf of the organisation</li>\n  <li>will sign the grant funding agreement if your application is successful</li>\n</ul>",
-                    "values": {
-                        "type": "listRef"
-                    },
-                    "schema": {}
-                }
-            ],
-            "next": [
-                {
-                    "path": "/summary"
-                },
-                {
-                    "path": "/authorised-signatory-details",
-                    "condition": "gbOJSj"
-                }
-            ],
-            "section": "roYQzW"
-        },
-        {
-            "title": "Summary",
-            "path": "/summary",
-            "controller": "./pages/summary.js",
-            "components": [],
-            "next": [],
-            "section": "roYQzW"
-        },
-        {
-            "path": "/authorised-signatory-details",
-            "title": "Authorised signatory details",
-            "components": [
-                {
-                    "name": "JTDVSR",
-                    "options": {},
-                    "type": "TextField",
-                    "title": "Authorised signatory full name",
-                    "schema": {}
-                },
-                {
-                    "name": "yafapW",
-                    "options": {},
-                    "type": "TextField",
-                    "title": "Authorised signatory job title",
-                    "schema": {}
-                },
-                {
-                    "type": "EmailAddressField",
-                    "title": "Authorised signatory email address",
-                    "name": "authorised-signatory-email",
-                    "nameHasError": false,
-                    "options": {
-                        "exposeToContext": false
-                    },
-                    "schema": {}
-                },
-                {
-                    "name": "CXrlUt",
-                    "options": {},
-                    "type": "TelephoneNumberField",
-                    "title": "Authorised signatory telephone number",
-                    "schema": {}
-                }
-            ],
-            "next": [
-                {
-                    "path": "/summary"
-                }
-            ],
-            "section": "roYQzW"
-        }
-    ],
-    "lists": [],
-    "sections": [
-        {
-            "name": "roYQzW",
-            "title": "Contact details",
-            "hideTitle": false
-        }
-    ],
-    "conditions": [
-        {
-            "displayName": "Authorised signatory different from lead contact",
-            "name": "gbOJSj",
-            "value": {
-                "name": "Authorised signatory different from lead contact",
-                "conditions": [
-                    {
-                        "field": {
-                            "name": "SDJIoS",
-                            "type": "YesNoField",
-                            "display": "Is the lead contact the same person as the authorised signatory?"
-                        },
-                        "operator": "is",
-                        "value": {
-                            "type": "Value",
-                            "value": "false",
-                            "display": "false"
-                        }
-                    }
-                ]
-            }
-        }
-    ],
-    "fees": [],
-    "outputs": [],
-    "version": 2,
-    "skipSummary": false,
-    "feeOptions": {
-        "allowSubmissionWithoutPayment": true,
-        "maxAttempts": 3,
-        "showPaymentSkippedWarningPage": false
-    },
-    "markAsComplete": false,
-    "name": "Contact details",
-    "feedback": {
-        "feedbackForm": false,
-        "url": ""
-    },
-    "phaseBanner": {}
-}
